refactor: 拆分 SimpleComponentProcessorFactory
类拆分:
- 抽出 ProcessorUtils 承载 processor/processorRequireSource/unsupported
  三个公共构建方法，供各工厂类复用
- enchantments/attributes/item_flag/lore 四个复杂组件各自独立为
  ComponentProcessorFactory 实现类，与 CustomPersistentDataProcessorFactory 同级
- SimpleComponentProcessorFactory 保留 14 个简单组件，678 行降至 385 行
- 各策略分支从匿名 lambda 提为命名方法，补充合并语义注释
  （如 enchantments 复制用 > 、合并用 >= 的非对称行为）
- ResultProcessorManager 显式注册新增的四个工厂
