+++ core/src/main/java/pers/yufiria/craftorithm/recipe/resultProcessor/ResultProcessorManager.java
@@ -4,7 +4,11 @@ import crypticlib.util.IOHelper;
 import org.bukkit.NamespacedKey;
 import org.bukkit.configuration.ConfigurationSection;
 import org.jetbrains.annotations.Nullable;
+import pers.yufiria.craftorithm.recipe.resultProcessor.impl.AttributesProcessorFactory;
 import pers.yufiria.craftorithm.recipe.resultProcessor.impl.CustomPersistentDataProcessorFactory;
+import pers.yufiria.craftorithm.recipe.resultProcessor.impl.EnchantmentsProcessorFactory;
+import pers.yufiria.craftorithm.recipe.resultProcessor.impl.ItemFlagProcessorFactory;
+import pers.yufiria.craftorithm.recipe.resultProcessor.impl.LoreProcessorFactory;
 import pers.yufiria.craftorithm.recipe.resultProcessor.impl.SimpleComponentProcessorFactory;
 
 import java.util.ArrayList;
@@ -28,6 +32,10 @@ public enum ResultProcessorManager {
             }
         }
         registerFactory(new CustomPersistentDataProcessorFactory());
+        registerFactory(EnchantmentsProcessorFactory.INSTANCE);
+        registerFactory(AttributesProcessorFactory.INSTANCE);
+        registerFactory(ItemFlagProcessorFactory.INSTANCE);
+        registerFactory(LoreProcessorFactory.INSTANCE);
     }
 
     public void registerFactory(ComponentProcessorFactory factory) {
